Repaired follow/unfollow/gotosettings button on profile screen
