+++ TrueApp/Storyboards/Profile/Home.storyboard
@@ -101,9 +101,6 @@
                                                             <integer key="value" value="3"/>
                                                         </userDefinedRuntimeAttribute>
                                                     </userDefinedRuntimeAttributes>
-                                                    <connections>
-                                                        <segue destination="avf-lz-Wyp" kind="show" id="A5i-cF-4p8"/>
-                                                    </connections>
                                                 </button>
                                                 <stackView opaque="NO" contentMode="scaleToFill" distribution="fillEqually" spacing="10" translatesAutoresizingMaskIntoConstraints="NO" id="ube-WI-48z">
                                                     <rect key="frame" x="0.0" y="8" width="167" height="29"/>
@@ -343,6 +340,9 @@
                                             <rect key="frame" x="8" y="519" width="86" height="29"/>
                                             <fontDescription key="fontDescription" type="system" pointSize="14"/>
                                             <state key="normal" title="Filter by Date"/>
+                                            <connections>
+                                                <action selector="filterByDateButtonTouched:" destination="brQ-EF-iUj" eventType="touchUpInside" id="wxa-zx-wsg"/>
+                                            </connections>
                                         </button>
                                     </subviews>
                                     <color key="backgroundColor" white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
@@ -1011,7 +1011,7 @@
         <!--Setting Table View Controller-->
         <scene sceneID="6sL-vX-PIK">
             <objects>
-                <tableViewController id="avf-lz-Wyp" customClass="SettingTableViewController" customModule="TrueApp" customModuleProvider="target" sceneMemberID="viewController">
+                <tableViewController storyboardIdentifier="SettingTableViewController" id="avf-lz-Wyp" customClass="SettingTableViewController" customModule="TrueApp" customModuleProvider="target" sceneMemberID="viewController">
                     <tableView key="view" clipsSubviews="YES" contentMode="scaleToFill" alwaysBounceVertical="YES" dataMode="static" style="grouped" separatorStyle="default" allowsSelection="NO" rowHeight="-1" estimatedRowHeight="-1" sectionHeaderHeight="18" sectionFooterHeight="18" id="Qee-SO-Hkj">
                         <rect key="frame" x="0.0" y="0.0" width="375" height="812"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
@@ -1526,7 +1526,4 @@
         <image name="home_icon" width="25" height="25"/>
         <image name="squareprofile" width="225" height="225"/>
     </resources>
-    <inferredMetricsTieBreakers>
-        <segue reference="A5i-cF-4p8"/>
-    </inferredMetricsTieBreakers>
 </document>
