+++ TrueApp/ViewControllers/Search/TrueSearchViewController.swift
@@ -91,7 +91,12 @@ extension TrueSearchViewController: UITableViewDelegate {
 }
 
 extension TrueSearchViewController: HeaderProfileCollectionReusableViewDelegate {
+    func goToSettingVC() {
+        
+    }
+    
     func updateFollowButton(forUser user: User) {
+        // TODO: Check is it work
         for u in self.users{
             if u.id == user.id{
                 u.followingSet = user.followingSet
